third_party/blink/web_tests/external/wpt/css/css-inline/parsing/baseline-shift-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['baseline-shift'] = "0" should set the property value
  assert_equals: serialization should be canonical expected "0px" but got "0"
[FAIL] e.style['baseline-shift'] = "bottom" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['baseline-shift'] = "center" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['baseline-shift'] = "top" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
